Fix group_id column type in activities table from bigInteger to integer

+++ database/migrations/2025_01_22_000003_fix_group_id_column_type_in_activities_table.php
@@ -28,14 +28,14 @@ return new class extends Migration
                 });
                 
                 Schema::table('activities', function (Blueprint $table) {
-                    $table->unsignedBigInteger('group_id')->nullable();
+                    $table->unsignedInteger('group_id')->nullable();
                     $table->foreign('group_id')->references('id')->on('groups')->onDelete('set null');
                 });
             }
         } else {
             // Column doesn't exist, create it
             Schema::table('activities', function (Blueprint $table) {
-                $table->unsignedBigInteger('group_id')->nullable();
+                $table->unsignedInteger('group_id')->nullable();
                 $table->foreign('group_id')->references('id')->on('groups')->onDelete('set null');
             });
         }
